pyproject.toml
[project]
name = "finance-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Arief Budhiman"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "cloudinary (>=1.44.0,<2.0.0)",
    "openai (>=1.73.0,<2.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "python-dotenv (>=0.21.0,<1.0.0)",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "langchain[openai]>=0.3.23",
    "langchain-community>=0.3.7",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
